ARREGLO DE DTO , Y CORRECCION DE POSTMAN URL QUE NO FUNCIONABA DE LISTAR ALMACENES POR ID
